+++ NEWS.md
@@ -1,7 +1,22 @@
-# AzureStor 3.4.1.9000
+# AzureStor 3.5.1.9000
 
-- Fix `storage_save_rds` and `storage_load_rds` to handle compression correctly. In particular, `storage_load_rds` should now correctly load files saved with `saveRDS`.
+- Add support for recursive deleting of non-empty subdirectories in blob and file storage.
+- Fix a bug when listing files in an empty ADLS filesystem (#100).
+
+# AzureStor 3.5.1
+
+- Fix for changed behaviour in readr 2.0.0, where `readr::read_delim` now automatically closes an input connection (#95).
+- Change `.onLoad` to keep any currently set global options when AzureStor is loaded, rather than resetting them (#96).
+
+# AzureStor 3.5.0
+
+- Add support for generating a service SAS. There is a new S3 generic `get_service_sas` with methods for `az_storage` and `storage_endpoint` objects, and a similar R6 method for `az_storage` objects. See `?sas` for more information.
+- Fix `storage_save_rds` and `storage_load_rds` to handle compression correctly. In particular, `storage_load_rds` should now correctly load files saved with `saveRDS` (#83).
 - Fix a bug that caused `list_blobs` to fail when leases were present.
+- Use a raw connection instead of a raw vector when calling `readr::read_delim` and `read_csv2`. This works around an issue introduced in readr 1.4.0 (#85, #86).
+- Update client API version to "2020-04-08". In particular, this allows specifying `resource_type="d"` when creating a service or user delegation SAS for blob storage.
+- Add an optional `service` argument to `storage_endpoint`, to specify the service in question: blob, file, ADLS2, queue or table. This allows use of the generic endpoint function with URLs that don't fit the usual pattern where the service is part of the hostname, eg custom domain names, IP addresses, etc.
+  - For the same reason, remove the warning about an unrecognised endpoint URL from `blob_endpoint`, `file_endpoint` and `adls_endpoint`.
 
 # AzureStor 3.4.1